extension: rm {Operator,User}Node
